[feature] #WCMS-4 - Create newsletter mod (#4)

+++ src/AppBundle/Controller/Admin/ProductController.php
@@ -2,10 +2,11 @@
 
 namespace AppBundle\Controller\Admin;
 
-use AppBundle\Entity\Photo;
+use AppBundle\Entity\Image;
 use AppBundle\Entity\Product;
-use AppBundle\Form\PhotoType;
-use AppBundle\Form\ProductType;
+use AppBundle\Form\Type\ImageType;
+use AppBundle\Form\Type\ProductDeleteType;
+use AppBundle\Form\Type\ProductType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -24,7 +25,9 @@ class ProductController extends Controller
      * @return Response
      */
     public function listAction() {
-        $products = $this->getDoctrine()->getRepository(Product::class)->findAll();
+        $products = $this->getDoctrine()
+            ->getRepository(Product::class)
+            ->findAll();
 
         return $this->render('admin/product/list.html.twig', [
             'products' => $products
@@ -47,9 +50,9 @@ class ProductController extends Controller
             $em->persist($product);
             $em->flush();
 
-            $this->addFlash('success', 'Produit ajouté avec succès.');
+            $this->addFlash('success', 'Product added successfully.');
 
-            return $this->redirectToRoute('admin_product_view', [
+            return $this->redirectToRoute('admin_product', [
                 'id' => $product->getId(),
             ]);
         }
@@ -67,8 +70,9 @@ class ProductController extends Controller
      * @return RedirectResponse|Response
      */
     public function editAction(Request $request, $id) {
-        $product = $this->getDoctrine()->getRepository(Product::class)->find($id);
-
+        $product = $this->getDoctrine()
+            ->getRepository(Product::class)
+            ->find($id);
         $this->checkProduct($product);
 
         $form = $this->createForm(ProductType::class, $product);
@@ -78,9 +82,9 @@ class ProductController extends Controller
             $em->persist($product);
             $em->flush();
 
-            $this->addFlash('success', 'Produit ajouté avec succès.');
+            $this->addFlash('success', 'Product edited successfully.');
 
-            return $this->redirectToRoute('admin_product_view', [
+            return $this->redirectToRoute('admin_product', [
                 'id' => $id,
             ]);
         }
@@ -93,31 +97,47 @@ class ProductController extends Controller
 
     /**
      * @Route("/{id}/delete", name="admin_product_delete", requirements={"id": "\d+"})
-     * @Method({"GET"})
+     * @Method({"GET", "POST"})
+     * @param $id
      * @param Request $request
      * @return RedirectResponse|Response
      */
-    public function deleteAction(Request $request, $id) {
-        $product = $this->getDoctrine()->getRepository(Product::class)->find($id);
-
+    public function deleteAction($id, Request $request) {
+        $product = $this->getDoctrine()
+            ->getRepository(Product::class)
+            ->find($id);
         $this->checkProduct($product);
 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($product);
-        $em->flush();
+        $form = $this->createForm(ProductDeleteType::class, $product);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($product);
+            $em->flush();
 
-        return $this->redirectToRoute('admin_products');
+            $this->addFlash('success', 'Newsletter deleted successfully.');
+
+            return $this->redirectToRoute('admin_products', [
+                'id' => $id,
+            ]);
+        }
+
+        return $this->render('admin/product/delete.html.twig', [
+            'form' => $form->createView(),
+            'product' => $product,
+        ]);
     }
 
     /**
-     * @Route("/{id}", name="admin_product_view", requirements={"id": "\d+"})
+     * @Route("/{id}", name="admin_product", requirements={"id": "\d+"})
      * @Method({"GET"})
      * @param integer $id
      * @return Response
      */
     public function viewAction($id) {
-        $product = $this->getDoctrine()->getRepository(Product::class)->find($id);
-
+        $product = $this->getDoctrine()
+            ->getRepository(Product::class)
+            ->find($id);
         $this->checkProduct($product);
 
         return $this->render('admin/product/view.html.twig', [
@@ -126,40 +146,43 @@ class ProductController extends Controller
     }
 
     /**
-     * @Route("/{id}/add_photo", name="admin_product_add_photo", requirements={"id": "\d+"})
+     * @Route("/{id}/add_image", name="admin_product_add_image", requirements={"id": "\d+"})
      * @Method({"GET", "POST"})
      * @param Request $request
      * @param integer $id
      * @return RedirectResponse|Response
      */
-    public function addPhotoAction(Request $request, $id) {
-        $product = $this->getDoctrine()->getRepository(Product::class)->find($id);
+    public function addImageAction(Request $request, $id) {
+        $product = $this->getDoctrine()
+            ->getRepository(Product::class)
+            ->find($id);
+        $this->checkProduct($product);
 
-        $photo = new Photo();
-        $photo->setProduct($product);
+        $image = new Image();
+        $image->setProduct($product);
 
-        $form = $this->createForm(PhotoType::class, $photo);
+        $form = $this->createForm(ImageType::class, $image);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            /** @var UploadedFile $path */
-            $path = $photo->getPath();
-            $file = md5(uniqid(null, true));
+            /** @var UploadedFile $file */
+            $file = $image->getPath();
+            $fileName = md5(uniqid(null, true));
             $filePath = $this->get('kernel')->getRootDir().'/../uploads/product/';
-            $path->move($filePath, $file);
-            $photo->setPath($file);
+            $file->move($filePath, $fileName);
+            $image->setPath($fileName);
 
             $em = $this->getDoctrine()->getManager();
-            $em->persist($photo);
+            $em->persist($image);
             $em->flush();
 
-            $this->addFlash('success_photo', 'Photo ajoutée avec succès.');
+            $this->addFlash('success_image', 'Image added successfully.');
 
-            return $this->redirectToRoute('admin_product_view', [
+            return $this->redirectToRoute('admin_product', [
                 'id' => $id
             ]);
         }
 
-        return $this->render('admin/product/modal/add_photo.html.twig', [
+        return $this->render('admin/product/modal/add_image.html.twig', [
             'form' => $form->createView(),
             'product' => $product,
         ]);
